pyproject.toml
[tool.poetry]
name = "pesquisa"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.2"}
ignr = "^2.2"
uvicorn = {extras = ["standard"], version = "^0.31.1"}
jinja2 = "^3.1.4"
sqlalchemy = "^2.0.35"
click = "^8.1.7"
pytz = "^2024.2"
alembic = "^1.13.3"
pydantic-settings = "^2.5.2"
jwt = "^1.3.1"
pwdlib = "^0.2.1"
pyotp = "^2.9.0"
qrcode = "^8.0"
psycopg = {extras = ["binary"], version = "^3.2.3"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.6.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
run = 'fastapi dev fast_zero/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fast_zero -vv'
post_test = 'coverage html'

